requirements.txt
argparse >= 1.4.0
PyYaml >= 3.11
logging >= 0.4.9.6
pytest >= 2.9.1
tox >= 2.3.1
click >= 6.6